correct lib_apqc_html.py path
